REFACTOR - updated files
Se movieron archivos de database del módulo de history al de calculator ya que ahí se usarán posteriormente.
Se agregó una fun más para el Dao conversion history para retornar flow o no.
Se renombraron archivos
