+++ app/src/main/java/com/varqulabs/dolarblue/calculator/data/local/database/dao/ConversionDao.kt
@@ -1,0 +1,12 @@
+package com.varqulabs.dolarblue.calculator.data.local.database.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import com.varqulabs.dolarblue.calculator.data.local.database.entities.ConversionEntity
+
+@Dao
+interface ConversionDao {
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertConversion(conversionEntity: ConversionEntity)
+}
