+++ app/src/main/java/com/varqulabs/dolarblue/calculator/data/local/database/entities/ConversionEntity.kt
@@ -1,4 +1,4 @@
-package com.varqulabs.dolarblue.history.data.local.database.entities
+package com.varqulabs.dolarblue.calculator.data.local.database.entities
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
@@ -6,11 +6,10 @@ import androidx.room.PrimaryKey
 import com.varqulabs.dolarblue.history.domain.model.Conversion
 import java.time.LocalDateTime
 
-// TODO: @Deivid - mover el entity al módulo de calculadora
 @Entity(tableName = "conversion_table")
 data class ConversionEntity(
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = "id") val id: Int,
+    @ColumnInfo(name = "id") val id: Int = 0,
     @ColumnInfo(name = "pesosBob") val pesosBob: String,
     @ColumnInfo(name = "pesosArg") val pesosArg: String,
     @ColumnInfo(name = "dollar") val dollar: String,
