--- app/src/main/java/com/varqulabs/dolarblue/history/data/local/database/dao/ConversionDao.kt
@@ -1,12 +1,0 @@
-package com.varqulabs.dolarblue.history.data.local.database.dao
-
-import androidx.room.Dao
-import androidx.room.Insert
-import com.varqulabs.dolarblue.history.data.local.database.entities.ConversionEntity
-
-// TODO @Deivid - mover el dao al módulo de calculadora
-@Dao
-interface ConversionDao {
-    @Insert()
-    suspend fun insertConversion(conversionEntity: ConversionEntity)
-}
