+++ app/src/main/java/com/varqulabs/dolarblue/history/data/local/database/entities/relations/ConversionsHistoryRelation.kt
@@ -2,9 +2,9 @@ package com.varqulabs.dolarblue.history.data.local.database.entities.relations
 
 import androidx.room.Embedded
 import androidx.room.Relation
-import com.varqulabs.dolarblue.history.data.local.database.entities.ConversionEntity
-import com.varqulabs.dolarblue.history.data.local.database.entities.CurrentExchangeRateEntity
-import com.varqulabs.dolarblue.history.data.local.database.entities.mapToModel
+import com.varqulabs.dolarblue.calculator.data.local.database.entities.ConversionEntity
+import com.varqulabs.dolarblue.calculator.data.local.database.entities.CurrentExchangeRateEntity
+import com.varqulabs.dolarblue.calculator.data.local.database.entities.mapToModel
 import com.varqulabs.dolarblue.history.domain.model.ConversionsHistory
 
 data class ConversionsHistoryRelation(
