+++ app/src/main/java/com/varqulabs/dolarblue/history/data/repository/ConversionsHistoryRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.varqulabs.dolarblue.history.data.repository
 
-import com.varqulabs.dolarblue.history.data.local.database.dao.ConversionHistoryDao
+import com.varqulabs.dolarblue.history.data.local.database.dao.ConversionsHistoryDao
 import com.varqulabs.dolarblue.history.data.local.database.entities.relations.mapToModel
 import com.varqulabs.dolarblue.history.domain.model.ConversionsHistory
 import com.varqulabs.dolarblue.history.domain.repository.ConversionsHistoryRepository
@@ -8,9 +8,9 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
 class ConversionsHistoryRepositoryImpl(
-    private val conversionHistoryDao: ConversionHistoryDao
+    private val conversionHistoryDao: ConversionsHistoryDao
 ): ConversionsHistoryRepository {
-    override suspend fun getConversionsHistory(): Flow<List<ConversionsHistory>> {
-        return conversionHistoryDao.getConversionsHistory().map { it.map { it.mapToModel() } }
+    override suspend fun getConversionsHistoryFlow(): Flow<List<ConversionsHistory>> {
+        return conversionHistoryDao.getConversionsHistoryFlow().map { it.map { it.mapToModel() } }
     }
 }
