+++ app/src/main/java/com/varqulabs/dolarblue/history/di/RepositoryModule.kt
@@ -1,9 +1,9 @@
 package com.varqulabs.dolarblue.history.di
 
-import com.varqulabs.dolarblue.history.data.local.database.dao.ConversionHistoryDao
+import com.varqulabs.dolarblue.history.data.local.database.dao.ConversionsHistoryDao
 import com.varqulabs.dolarblue.history.data.repository.ConversionsHistoryRepositoryImpl
 import com.varqulabs.dolarblue.history.domain.repository.ConversionsHistoryRepository
-import com.varqulabs.dolarblue.history.domain.usecases.GetConversionsHistoryUseCase
+import com.varqulabs.dolarblue.history.domain.usecases.GetConversionsHistoryFlowUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -16,13 +16,13 @@ object RepositoryModule {
 
     @Provides
     @Singleton
-    fun provideConversionsHistoryRepository(conversionHistoryDao: ConversionHistoryDao) : ConversionsHistoryRepository {
-        return ConversionsHistoryRepositoryImpl(conversionHistoryDao)
+    fun provideConversionsHistoryRepository(conversionsHistoryDao: ConversionsHistoryDao) : ConversionsHistoryRepository {
+        return ConversionsHistoryRepositoryImpl(conversionsHistoryDao)
     }
 
     @Provides
     @Singleton
-    fun provideGetConversionUseCase(repository: ConversionsHistoryRepository): GetConversionsHistoryUseCase {
-        return GetConversionsHistoryUseCase(repository)
+    fun provideGetConversionsHistoryFlowUseCase(repository: ConversionsHistoryRepository): GetConversionsHistoryFlowUseCase {
+        return GetConversionsHistoryFlowUseCase(repository)
     }
 }
