+++ app/src/main/java/com/varqulabs/dolarblue/history/domain/repository/ConversionsHistoryRepository.kt
@@ -4,5 +4,5 @@ import com.varqulabs.dolarblue.history.domain.model.ConversionsHistory
 import kotlinx.coroutines.flow.Flow
 
 interface ConversionsHistoryRepository {
-    suspend fun getConversionsHistory(): Flow<List<ConversionsHistory>>
+    suspend fun getConversionsHistoryFlow(): Flow<List<ConversionsHistory>>
 }
