+++ app/src/main/java/com/varqulabs/dolarblue/history/domain/usecases/GetConversionsHistoryFlowUseCase.kt
@@ -4,10 +4,10 @@ import com.varqulabs.dolarblue.history.domain.model.ConversionsHistory
 import com.varqulabs.dolarblue.history.domain.repository.ConversionsHistoryRepository
 import kotlinx.coroutines.flow.Flow
 
-class GetConversionsHistoryUseCase(
+class GetConversionsHistoryFlowUseCase(
     private val conversionsHistoryRepository: ConversionsHistoryRepository
 ) {
     suspend operator fun invoke(): Flow<List<ConversionsHistory>> {
-        return conversionsHistoryRepository.getConversionsHistory()
+        return conversionsHistoryRepository.getConversionsHistoryFlow()
     }
 }
